move chunkArray tests
